# Address UnionInPlace performance regressions