## Overview

UnionInPlace was making a lot of copies and/or new bitmaps to simplify the merge logic. The logic for just grabbing the one container for a key was slightly broken by late changes to the UnionInPlace PR to simplify/clean up the logic to treat the target as a method receiver, which made the SummaryStats computation of whether a container was the "only" container slightly wrong. Thus, in the case where the receiver had no container, but one of the things being merged in did, we were creating a new container and unioning everything into it -- even if there was only one evereything to union.

Cleaned that up, also stopped coercing everything to bitmap right away. We coerce to bitmap if we need to, but special-case array/array merges to allow them to stay arrays if N stays small.

Fixes #1875

## Pull request checklist

- [ ] I have read the [contributing guide](https://github.com/pilosa/pilosa/blob/master/CONTRIBUTING.md).
- [ ] I have agreed to the [Contributor License Agreement](https://cla-assistant.io/pilosa/pilosa).
- [ ] I have updated the [documentation](https://github.com/pilosa/pilosa/tree/master/docs).
- [ ] I have resolved any merge conflicts.
- [ ] I have included tests that cover my changes.
- [ ] All new and existing tests pass.

## Code review checklist
This is the checklist that the reviewer will follow while reviewing your pull request. You do not need to do anything with this checklist, but be aware of what the reviewer will be looking for.

- [ ] Ensure that any changes to external docs have been included in this pull request.
- [ ] If the changes require that minor/major versions need to be updated, tag the PR appropriately.
- [ ] Ensure the new code is [properly commented](https://github.com/golang/go/wiki/CodeReviewComments#doc-comments) and follows [Idiomatic Go](https://dmitri.shuralyov.com/idiomatic-go).
- [ ] Check that tests have been written and that they cover the new functionality.
- [ ] Run tests and ensure they pass.
- [ ] Build and run the code, performing any applicable integration testing.