@jaffee 2019-03-13
I put this into fragment.ImportRoaring like: 
```
@@ -1763,8 +1763,11 @@ func (f *fragment) importRoaring(data []byte, clear bool) error {
 
 	if clear {
 		bm = f.storage.Difference(bm)
-	} else if f.storage.Any() {
-		bm = f.storage.Union(bm)
+	} else if f.storage.Containers.Size() >= bm.Containers.Size() {
+		f.storage.UnionInPlace(bm)
+		bm = f.storage
+	} else {
+		bm.UnionInPlace(f.storage)
 	}
 
 	for rowID := range rowSet {
```

Benchcmp looks good:
```
benchmark                                                     old ns/op     new ns/op     delta
BenchmarkImportRoaring/Rows2Cache_ranked-8                    6180105       6131043       -0.79%
BenchmarkImportRoaring/Rows50Cache_ranked-8                   17139516      16956038      -1.07%
BenchmarkImportRoaring/Rows1000Cache_ranked-8                 27238809      28194480      +3.51%
BenchmarkImportRoaring/Rows100000Cache_ranked-8               104570846     104415830     -0.15%
BenchmarkImportRoaringUpdate/ranked2Rows20Cols-8              524057        518185        -1.12%
BenchmarkImportRoaringUpdate/ranked2Rows1000Cols-8            542770        567699        +4.59%
BenchmarkImportRoaringUpdate/ranked2Rows50000Cols-8           7025988       6822882       -2.89%
BenchmarkImportRoaringUpdate/ranked2Rows500000Cols-8          7788524       6856893       -11.96%
BenchmarkImportRoaringUpdate/ranked50Rows20Cols-8             647889        690404        +6.56%
BenchmarkImportRoaringUpdate/ranked50Rows1000Cols-8           3140017       3125205       -0.47%
BenchmarkImportRoaringUpdate/ranked50Rows50000Cols-8          32562283      24226950      -25.60%
BenchmarkImportRoaringUpdate/ranked50Rows500000Cols-8         45576973      33472502      -26.56%
BenchmarkImportRoaringUpdate/ranked1000Rows20Cols-8           3167222       3147993       -0.61%
BenchmarkImportRoaringUpdate/ranked1000Rows1000Cols-8         3740793       3765449       +0.66%
BenchmarkImportRoaringUpdate/ranked1000Rows50000Cols-8        45418052      35601999      -21.61%
BenchmarkImportRoaringUpdate/ranked1000Rows500000Cols-8       63076642      49400175      -21.68%
BenchmarkImportRoaringUpdate/ranked100000Rows20Cols-8         4234861       4128218       -2.52%
BenchmarkImportRoaringUpdate/ranked100000Rows1000Cols-8       4965404       4858865       -2.15%
BenchmarkImportRoaringUpdate/ranked100000Rows50000Cols-8      117548306     99915298      -15.00%
BenchmarkImportRoaringUpdate/ranked100000Rows500000Cols-8     160903920     158869631     -1.26%
BenchmarkImportRoaringIntoLargeFragment-8                     497445827     395656499     -20.46%

benchmark                                                     old allocs     new allocs     delta
BenchmarkImportRoaring/Rows2Cache_ranked-8                    130            130            +0.00%
BenchmarkImportRoaring/Rows50Cache_ranked-8                   1746           1746           +0.00%
BenchmarkImportRoaring/Rows1000Cache_ranked-8                 33294          33293          -0.00%
BenchmarkImportRoaring/Rows100000Cache_ranked-8               256294         256308         +0.01%
BenchmarkImportRoaringUpdate/ranked2Rows20Cols-8              58             58             +0.00%
BenchmarkImportRoaringUpdate/ranked2Rows1000Cols-8            105            105            +0.00%
BenchmarkImportRoaringUpdate/ranked2Rows50000Cols-8           208            166            -20.19%
BenchmarkImportRoaringUpdate/ranked2Rows500000Cols-8          208            166            -20.19%
BenchmarkImportRoaringUpdate/ranked50Rows20Cols-8             130            130            +0.00%
BenchmarkImportRoaringUpdate/ranked50Rows1000Cols-8           1815           1815           +0.00%
BenchmarkImportRoaringUpdate/ranked50Rows50000Cols-8          10881          2596           -76.14%
BenchmarkImportRoaringUpdate/ranked50Rows500000Cols-8         11848          2595           -78.10%
BenchmarkImportRoaringUpdate/ranked1000Rows20Cols-8           133            133            +0.00%
BenchmarkImportRoaringUpdate/ranked1000Rows1000Cols-8         3471           3471           +0.00%
BenchmarkImportRoaringUpdate/ranked1000Rows50000Cols-8        100474         38935          -61.25%
BenchmarkImportRoaringUpdate/ranked1000Rows500000Cols-8       125769         50092          -60.17%
BenchmarkImportRoaringUpdate/ranked100000Rows20Cols-8         138            138            +0.00%
BenchmarkImportRoaringUpdate/ranked100000Rows1000Cols-8       3905           3905           +0.00%
BenchmarkImportRoaringUpdate/ranked100000Rows50000Cols-8      402957         162715         -59.62%
BenchmarkImportRoaringUpdate/ranked100000Rows500000Cols-8     574944         351759         -38.82%
BenchmarkImportRoaringIntoLargeFragment-8                     1567068        526777         -66.38%

benchmark                                                     old bytes     new bytes     delta
BenchmarkImportRoaring/Rows2Cache_ranked-8                    19331         19332         +0.01%
BenchmarkImportRoaring/Rows50Cache_ranked-8                   240952        240964        +0.00%
BenchmarkImportRoaring/Rows1000Cache_ranked-8                 5166511       5165697       -0.02%
BenchmarkImportRoaring/Rows100000Cache_ranked-8               40269989      40270966      +0.00%
BenchmarkImportRoaringUpdate/ranked2Rows20Cols-8              76712         76712         +0.00%
BenchmarkImportRoaringUpdate/ranked2Rows1000Cols-8            286856        286856        +0.00%
BenchmarkImportRoaringUpdate/ranked2Rows50000Cols-8           285213        282013        -1.12%
BenchmarkImportRoaringUpdate/ranked2Rows500000Cols-8          285287        281820        -1.22%
BenchmarkImportRoaringUpdate/ranked50Rows20Cols-8             141798        141793        -0.00%
BenchmarkImportRoaringUpdate/ranked50Rows1000Cols-8           3981197       3981218       +0.00%
BenchmarkImportRoaringUpdate/ranked50Rows50000Cols-8          7279204       2601867       -64.26%
BenchmarkImportRoaringUpdate/ranked50Rows500000Cols-8         11476959      3536559       -69.19%
BenchmarkImportRoaringUpdate/ranked1000Rows20Cols-8           50345         50341         -0.01%
BenchmarkImportRoaringUpdate/ranked1000Rows1000Cols-8         1702708       1702731       +0.00%
BenchmarkImportRoaringUpdate/ranked1000Rows50000Cols-8        12567697      6307264       -49.81%
BenchmarkImportRoaringUpdate/ranked1000Rows500000Cols-8       16307988      8458899       -48.13%
BenchmarkImportRoaringUpdate/ranked100000Rows20Cols-8         546803        546789        -0.00%
BenchmarkImportRoaringUpdate/ranked100000Rows1000Cols-8       2064089       2064080       -0.00%
BenchmarkImportRoaringUpdate/ranked100000Rows50000Cols-8      44150201      23329376      -47.16%
BenchmarkImportRoaringUpdate/ranked100000Rows500000Cols-8     64531238      41585229      -35.56%
BenchmarkImportRoaringIntoLargeFragment-8                     202440744     88946368      -56.06%
```