@seebs 2019-03-13
@richardartoul, can you have a look at this change? I am not 100% confident that I understand this code well enough, and I might have missed something relevant. But basically, there were performance regressions in the case where we were merging very sparse things (either two array containers, or an array container merged with no container), and i think this fixes them, but I'd like second opinions on it.