@richardartoul 2019-03-13
@seebs @jaffee Whoa this looks promising. Yeah I didn't have time to do this at the time, but looks like its a huge win!

I'll take a look through today or tomorrow if that works for you.

Also hopefully I didn't cause you any issues with the regression? I think the way I wrote the code is it still uses the old logic and you had to opt-in to the new path.